requirements.txt
pyyaml>=6.0.1
rich>=13.7.0
click>=8.1.7
requests>=2.31.0
python-whois>=0.9.4
